+++ iOS-MyCarMaster/MyCarMaster/MyCarMaster/Scene/Main/StepBottom/StepBottomView.swift
@@ -10,26 +10,6 @@ import UIKit
 
 import MVIFoundation
 
-final class GeneralButton: UIButton, ButtonStyleSeletable {
-
-    override var intrinsicContentSize: CGSize {
-        return CGSize(width: 0, height: 44)
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: .zero)
-        configureUI()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    private func configureUI() {
-        style = .titleMedium2(nil)
-    }
-}
-
 final class StepBottomView: UIView {
 
     override var intrinsicContentSize: CGSize {
@@ -68,17 +48,13 @@ final class StepBottomView: UIView {
         return label
     }()
 
-    let leftButton: GeneralButton = {
-        let button = GeneralButton()
-        button.unselectedStyle()
-        return button
-    }()
+    let leftButton = ThemeButton().then { button in
+        button.secondaryStyle()
+    }
 
-    let rightButton: GeneralButton = {
-        let button = GeneralButton()
+    let rightButton = ThemeButton().then { button in
         button.selectedStyle()
-        return button
-    }()
+    }
 
     let summaryButton: UIButton = {
         let button = UIButton()
